Set the number of CPUs to use
It seems like we're compiling slow on CI because the CPUs don't match.
The jobs run on servers with 32 CPUs, but are given 2 vCPUs.
When `purs` runs, it tries to maximize how many CPUs are used.
It grabs the number of CPUs on the machine, not available to it.
Ends up spending more time scheduling for 32 CPUs,
but not being able to use 32 CPUs.